+++ GISLibrary/Record.cpp
@@ -44,8 +44,8 @@ void Record::SetLeader(int totalFieldSize, bool adjustEntryMap)
 		}
 	}
 
-	leader.sizeOfFieldLengthField = LatLonUtility::countDigits(maxFieldLength);
-	leader.sizeOfFieldPositionField = LatLonUtility::countDigits(maxFieldPosition);
+	leader.sizeOfFieldLengthField = (int)LatLonUtility::countDigits(maxFieldLength);
+	leader.sizeOfFieldPositionField = (int)LatLonUtility::countDigits(maxFieldPosition);
 
 	if (adjustEntryMap)
 	{
@@ -73,5 +73,5 @@ void Record::SetLeader(int totalFieldSize, bool adjustEntryMap)
 
 int Record::DirectoryLength(int sizeOfFieldLengthField, int sizeOfFieldPositionField)
 {
-	return directory.size() * (4 + sizeOfFieldLengthField + sizeOfFieldPositionField) + 1;
+	return (int)(directory.size() * (4 + sizeOfFieldLengthField + sizeOfFieldPositionField) + 1);
 }
